% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon_fns.R
\name{anon_date_shift}
\alias{anon_date_shift}
\title{Shift dates to a new time period}
\usage{
anon_date_shift(x, center_date = Sys.Date(), scramble = FALSE)
}
\arguments{
\item{x}{A Date or POSIXct vector to anonymize}

\item{center_date}{New center point for the date range (default: current date)}

\item{scramble}{Logical, whether to randomize the assignment order of generated results}
}
\description{
Shifts dates to a new time period while preserving relative relationships
between dates. Supports Date and POSIXct objects.
}
\examples{
# Date shifting
dates <- as.Date(c("2023-01-15", "2023-02-20", "2023-03-10"))
anon_date_shift(dates, center_date = "2024-06-01")

}
