% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon_fns.R
\name{anon_email}
\alias{anon_email}
\title{Anonymize email addresses}
\usage{
anon_email(x, start = "user", domain = "domain.com")
}
\arguments{
\item{x}{A vector to anonymize}

\item{start}{Character prefix for generated results (default: "user")}

\item{domain}{Domain for generated emails (default: "domain.com")}
}
\description{
Anonymizes email addresses by replacing them with sequential fake emails.
}
\examples{
# Email anonymization
emails <- c("john@company.com", "not_an_email", "jane at work dot org")
anon_email(emails)

}
