% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon_fns.R
\name{anon_id_num_sequence}
\alias{anon_id_num_sequence}
\title{Convert unique identifiers to sequential numeric IDs}
\usage{
anon_id_num_sequence(x, scramble = FALSE)
}
\arguments{
\item{x}{A vector to anonymize}

\item{scramble}{Logical, whether to randomize the assignment order of generated results}
}
\description{
Convert unique identifiers to sequential numeric IDs
}
