% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon_fns.R
\name{anon_num_preserve_distribution}
\alias{anon_num_preserve_distribution}
\title{Anonymize numeric data while preserving distribution}
\usage{
anon_num_preserve_distribution(
  x,
  method = c("rank", "noise", "quantile"),
  noise_sd = NULL,
  quantile_dist_family = c("normal", "uniform", "exponential")
)
}
\arguments{
\item{x}{A numeric vector to anonymize}

\item{method}{Transformation method: "rank", "noise", or "quantile"}

\item{noise_sd}{Standard deviation for noise method (used when method = "noise") (default: NULL resulting in standard deviation being calculated)}

\item{quantile_dist_family}{Distribution family for quantile method: "normal", "uniform", "exponential" (used when method = "quantile")}
}
\description{
Anonymizes numeric data while preserving distributional properties using
various transformation methods.
}
