% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon_fns.R
\name{anon_phone_number}
\alias{anon_phone_number}
\title{Anonymize phone numbers}
\usage{
anon_phone_number(x, start = "555-000")
}
\arguments{
\item{x}{A character vector to anonymize}

\item{start}{Character prefix for generated results (default: "555-000")}
}
\description{
Anonymizes phone numbers by replacing them with sequential fake numbers.
}
\examples{
phones <- c("Call me at 123-456-7890", "1234567890", "(123) 456-7890")
anon_phone_number(phones)

}
