% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mapper.R
\name{as_function_formula}
\alias{as_function_formula}
\title{Convert a formula to a function in a manner compatible with \code{rlang::as_function()} but with more features.}
\usage{
as_function_formula(form, .env = parent.frame())
}
\arguments{
\item{form}{formula. one- or two-sided}

\item{.env}{environment of function}
}
\value{
function derived from the formula
}
\description{
This is a modified form of \code{anon::lambda()} which only accepts a single
argument which must be a formula (either one- or two-sides).
}
\details{
Use this as an alternate \code{as_mapper} in \code{purrr} by running \code{anon::patch_purrr}.
}
