% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.r
\name{cell_swap}
\alias{cell_swap}
\title{Randomly shuffle the values in each column of a data frame.}
\usage{
cell_swap(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A new data frame with the same dimensions as data, but with
        the values in each column randomly shuffled.
}
\description{
This function takes a data frame as input and returns a new data frame
with the same dimensions as the original data, but with the values in each
column randomly shuffled.
}
\examples{
data <- data.frame(x = c(1, 2, 3), y = c(4, 5, 6))
shuffled_data <- cell_swap(data)
shuffled_data

}
