% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{coeff_picker_list}
\alias{coeff_picker_list}
\title{Coefficient Picker List}
\usage{
coeff_picker_list(multimodel_lm_list, interval, intercept)
}
\arguments{
\item{multimodel_lm_list}{A list of models where each element is a list containing a reference model (non-anonymous) and an anonymous model.}

\item{interval}{The confidence interval level.}

\item{intercept}{Logical indicating whether to include the intercept in the calculation.}
}
\value{
Returns a list containing information about the means and confidence intervals of coefficients for each model.
}
\description{
A function to calculate the mean of the means and the mean of the confidence intervals for a list of models.
}
