% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{coeff_plot_list}
\alias{coeff_plot_list}
\title{Coefficient Plot List}
\usage{
coeff_plot_list(
  multimodel_lm_list,
  model_names = NULL,
  interval = 0.95,
  intercept = FALSE
)
}
\arguments{
\item{multimodel_lm_list}{A list of models where each element is a list containing a reference model (non-anonymous) and an anonymous model.}

\item{model_names}{Optional. Names for the models. If not provided, default names ("Model1", "Model2", etc.) will be used.}

\item{interval}{The confidence interval level.}

\item{intercept}{Logical indicating whether to include the intercept in the plot.}
}
\description{
A function to create a plot of coefficients with point estimates and confidence intervals for a list of models.
}
