% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{colSds}
\alias{colSds}
\title{Calculate Column standard deviations}
\usage{
colSds(df, na.rm = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{na.rm}{Logical. Should missing values (NA) be removed?}
}
\value{
A numeric vector of standard deviations for each column.
}
\description{
Calculate the standard deviations of columns in a data frame.
}
\examples{
\dontrun{
variances <- ColSds(data_frame)
}
}
