% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{column_inserter}
\alias{column_inserter}
\title{Column Inserter}
\usage{
column_inserter(data, column, colname, index)
}
\arguments{
\item{data}{The data frame to insert the column into.}

\item{column}{The column to insert.}

\item{colname}{The name of the column to insert.}

\item{index}{The index at which to insert the column.}
}
\value{
The data frame with the new column inserted at the specified index.
}
\description{
Insert a column into a data frame at a specified index.
}
