% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{column_reordered}
\alias{column_reordered}
\title{Column Reordered}
\usage{
column_reordered(data, names, indicies)
}
\arguments{
\item{data}{The data frame to reorder.}

\item{names}{The column names in the desired order.}

\item{indicies}{The target indices for the corresponding column names.}
}
\value{
The data frame with columns reordered as specified.
}
\description{
Reorder the data frame columns according to the specified column names and indices.
}
