% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{colVars}
\alias{colVars}
\title{Calculate Column Variances}
\usage{
colVars(df, na.rm = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{na.rm}{Logical. Should missing values (NA) be removed?}
}
\value{
A numeric vector of variances for each column.
}
\description{
Calculate the variances of columns in a data frame.
}
\examples{
\dontrun{
variances <- colVars(data_frame)
}
}
