% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_non_ordered_classes.r
\name{combine_lowest_classes}
\alias{combine_lowest_classes}
\title{Combine the lowest classes}
\usage{
combine_lowest_classes(column, k = 1)
}
\arguments{
\item{column}{The column in which to combine the lowest classes.}

\item{k}{The maximum number of lowest classes to combine.}
}
\value{
The column with the lowest classes combined.
}
\description{
This function combines the lowest classes in a column up to a specified number (k).
The lowest classes are determined based on the frequency of occurrence in the column.
}
\examples{
column <- c("A", "B", "A", "C", "B", "D", "A", "C", "B")
combine_lowest_classes(column, 2)
}
