% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon_measures.r
\name{distances}
\alias{distances}
\title{Calculate the minimum distance between rows of two datasets.}
\usage{
distances(original_data, reference_data = NULL, dist = euc_dist)
}
\arguments{
\item{original_data}{A data frame representing the original dataset.}

\item{reference_data}{A data frame representing the reference dataset.}

\item{dist}{A function that calculates the distance between two rows of a data frame.}
}
\value{
A numeric vector containing the minimum distances between rows of the
        \code{original_data} and \code{reference_data}.
}
\description{
This function calculates the minimum distance between each row of the
\code{original_data} and every row of the \code{reference_data} using a specified
distance function (\code{dist}). The datasets are first one-hot encoded to enable
distance calculations. The function then scales the datasets to have mean 0 and
standard deviation 1. If \code{cross_val} is set to \code{TRUE}, the function
performs leave-one-out cross-validation to calculate the distance between each row
of the \code{original_data} against the rest of the \code{original_data}. If
\code{cross_val} is \code{FALSE}, the function calculates the distance between
each row of the \code{original_data} and every row of the \code{reference_data}.
}
