% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{emptydf}
\alias{emptydf}
\title{Create an empty data frame with the same dimensions and column names as the input data.}
\usage{
emptydf(data)
}
\arguments{
\item{data}{A data frame to replicate the structure from.}
}
\value{
An empty data frame with the same dimensions and column names as \code{data}.
}
\description{
This function takes a data frame as input and returns a new empty data frame
with the same dimensions and column names as the input data. The returned data frame
will have no data but can be populated later with relevant information.
}
