% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto.r
\name{encrypt}
\alias{encrypt}
\title{Encrypts the data using RSA encryption and scales the encrypted values to [0, 1].}
\usage{
encrypt(data, my_key = FALSE, bits = 2048)
}
\arguments{
\item{data}{The data frame to be encrypted and scaled.}

\item{my_key}{Logical parameter: whether your own key from .ssh folder should be used or not}

\item{bits}{The number of bits used for precision in the RSA encryption.}
}
\value{
A scaled data frame where each column contains the encrypted and scaled values.
}
\description{
This function takes a data frame as input, encrypts each column using RSA encryption
with the provided number of bits and scales the encrypted values to the [0, 1] range.
}
\examples{
data <- data.frame(x = c(10, 20, 30), y = c(5, 15, 25))
encrypted_data <- encrypt(data, bits = 2048)
print(encrypted_data)

}
