% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy.r
\name{euc_dist}
\alias{euc_dist}
\title{Calculate the Euclidean distance between two vectors normalized by vector length.}
\usage{
euc_dist(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as \code{x}.}
}
\value{
The Euclidean distance between \code{x} and \code{y} normalized by vector length.
}
\description{
This function calculates the Euclidean distance between two vectors \code{x} and \code{y},
which are normalized by the length of the vectors. The Euclidean distance is a measure
of the straight-line distance between two points in a space. Normalizing the distance
by the vector length ensures that the magnitude of the distance is independent of the
vector length, allowing for fair comparisons between vectors of different lengths.
}
\examples{
x <- c(1, 2, 3)
y <- c(4, 5, 6)
distance <- euc_dist(x, y)
distance

}
