% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_div_subsets.r
\name{findNearestSubset}
\alias{findNearestSubset}
\title{Find the index of the nearest subset}
\usage{
findNearestSubset(subset, subsets, quasiIdentifiers)
}
\arguments{
\item{subset}{The subset for which to find the nearest subset.}

\item{subsets}{A list of subsets to compare with the given subset.}

\item{quasiIdentifiers}{A character vector specifying the quasi-identifier columns.}
}
\value{
The index of the nearest subset in the list.
}
\description{
This function finds the index of the nearest subset to a given subset among a list of subsets, based on the specified quasi-identifier columns.
}
\examples{
subset <- list(data.frame(
  Q1 = c("A", "B"),
  Q2 = c("X", "Y")
))
subsets <- list(
  data.frame(Q1 = "A", Q2 = "X"),
  data.frame(Q1 = "A", Q2 = "Y"),
  data.frame(Q1 = "B", Q2 = "X")
)
findNearestSubset(subset, subsets, c("Q1", "Q2"))

}
