% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-to-function.R
\name{formula_to_1arg_function}
\alias{formula_to_1arg_function}
\title{Conversion of formula to a 1-arg function}
\usage{
formula_to_1arg_function(form, .env = parent.frame())
}
\arguments{
\item{form}{formula}

\item{.env}{environment of function}
}
\value{
function
}
\description{
This aims to be a dependency free, very simple formula-to-function convertor.
}
\details{
The only supported argument is \code{.x}, all other variables are assumed to
come from the environment.
}
