% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-to-function.R
\name{formula_to_function}
\alias{formula_to_function}
\title{Convert a formula to a function}
\usage{
formula_to_function(form, .env = parent.frame())
}
\arguments{
\item{form}{formula}

\item{.env}{environment of function}
}
\value{
function whose code is taken from the RHS of the given formula, with
formal arguments of \code{.x, .y, .z}
}
\description{
This aims to be a dependency free, very simple formula-to-function convertor.
}
\details{
The only supported arguments are \code{.x, .y, .z}, all other variables are assumed to
come from the environment.
}
