% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{get_mode}
\alias{get_mode}
\title{Get Mode}
\usage{
get_mode(x)
}
\arguments{
\item{x}{The input vector.}
}
\value{
The mode of the input vector.
}
\description{
Find the mode (most frequent value) in a vector.
}
\examples{
data <- c(1, 2, 2, 3, 3, 3, 4, 4, 4, 4)
get_mode(data)

}
