% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{getMinMax}
\alias{getMinMax}
\title{Get the minimum and maximum values of a numeric vector}
\usage{
getMinMax(numbers)
}
\arguments{
\item{numbers}{A numeric vector.}
}
\value{
A character string representing the minimum and maximum values in the format [min, max].
}
\description{
This function takes a numeric vector as input and returns the minimum and maximum values.
}
\examples{
numbers <- c(1, 2, 3)
getMinMax(numbers)

}
