% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{inverse_logit}
\alias{inverse_logit}
\title{Inverse Logit Function}
\usage{
inverse_logit(x)
}
\arguments{
\item{x}{The input value.}
}
\value{
The result of the inverse logit function.
}
\description{
Calculate the inverse logit function.
}
\examples{
inverse_logit(0)
inverse_logit(-2)
inverse_logit(2)
}
