% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{inverse_onehot}
\alias{inverse_onehot}
\title{Inverse One-Hot Encoding}
\usage{
inverse_onehot(data, names, sample = FALSE)
}
\arguments{
\item{data}{The one-hot encoded dataset.}

\item{names}{The names of the categorical variables in the original data.
Names in the original are assumed to be mapped to "name=level"
in the one hot encoded data given. Works well with
"onehot" package.}

\item{sample}{Logical: Whether to sample from the numerical columns or take the max}
}
\value{
The original data with categorical variables restored from one-hot encoding.
}
\description{
Perform the inverse operation to one-hot encoding, i.e.,
 from one-hot encoded data to the original data.
}
