% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_anon_subsets.r
\name{is_k_anonymous}
\alias{is_k_anonymous}
\title{Check if a data frame is k-anonymous with respect to specified quasi-identifier columns}
\usage{
is_k_anonymous(data, quasi_id_cols, k)
}
\arguments{
\item{data}{The input data frame.}

\item{quasi_id_cols}{A character vector specifying the names of the quasi-identifier columns.}

\item{k}{The desired level of k-anonymity.}
}
\value{
TRUE if the data frame is k-anonymous, FALSE otherwise.
}
\description{
This function takes a data frame and checks if it satisfies k-anonymity with respect
to the specified quasi-identifier columns. It groups the data by the quasi-identifiers,
counts the number of rows in each group, and checks if all groups have at least k rows.
}
