% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_div_columnwise.r
\name{is_l_diverse}
\alias{is_l_diverse}
\title{Check if a data frame is l-diverse with respect to specified quasi-identifier and sensitive columns}
\usage{
is_l_diverse(df, quasi_id_cols, sensitive_cols, l)
}
\arguments{
\item{df}{The input data frame.}

\item{quasi_id_cols}{A character vector specifying the names of the quasi-identifier columns.}

\item{sensitive_cols}{A character vector specifying the names of the sensitive columns.}

\item{l}{The desired minimum number of distinct values for each sensitive column within each group.}
}
\value{
TRUE if the data frame is l-diverse, FALSE otherwise.
}
\description{
This function takes a data frame and checks if it satisfies l-diversity with respect to the
specified quasi-identifier and sensitive columns. It groups the data by the quasi-identifier
columns, counts the number of distinct values in each group for the sensitive columns, and
checks if all groups have at least l distinct values for each sensitive column.
}
\examples{
df <- data.frame(
  age = c(25, 30, 35, 40, 45),
  gender = c("M", "M", "F", "F", "M"),
  disease = c("A", "A", "B", "C", "C")
)

is_l_diverse(df, c("age", "gender"), "disease", 2)


}
