% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_div_subsets.r
\name{isLdiverse}
\alias{isLdiverse}
\title{Check if a DataFrame is l-diverse}
\usage{
isLdiverse(df, quasi_id_cols, sensitive_cols, l)
}
\arguments{
\item{df}{The DataFrame to be checked.}

\item{quasi_id_cols}{A character vector specifying the quasi-identifier columns.}

\item{sensitive_cols}{A character vector specifying the sensitive columns.}

\item{l}{The minimum number of rows required in each group.}
}
\value{
A logical value indicating whether the DataFrame is l-diverse.
}
\description{
This function checks if a DataFrame is l-diverse based on the specified quasi-identifier columns, sensitive columns, and l value.
}
\examples{
\dontrun{
df <- data.frame(
  Q1 = c("A", "A", "B", "B"),
  Q2 = c("X", "Y", "X", "Y"),
  S = c(1, 2, 3, 4)
)
isLdiverse(df, c("Q1", "Q2"), c("S"), 2)
}
}
