% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_anon_subsets.r
\name{kAnon_legacy}
\alias{kAnon_legacy}
\title{Perform k-Anonymization on a dataset}
\usage{
kAnon_legacy(data, quasiIdentifiers, diversityFunctions, k)
}
\arguments{
\item{data}{A data frame containing the sensitive data.}

\item{quasiIdentifiers}{A character vector specifying the column names of the quasi-identifiers.}

\item{diversityFunctions}{A named list of diversity functions for each quasi-identifier column.}

\item{k}{The desired level of k-anonymity.}
}
\value{
A k-anonymous data frame.
}
\description{
This function applies k-anonymization to a dataset by dividing it into subsets
based on quasi-identifiers and modifying the data to achieve k-anonymity
}
\details{
The `kAnon` function applies k-anonymization on the input dataset `data`
by dividing it into subsets based on quasi-identifiers specified in `quasiIdentifiers`.
The function modifies the data in each subset to achieve k-anonymity
based on the diversity functions provided in `diversityFunctions`.
}
\examples{
\dontrun{
data(iris)
diversityFunctions <- list(Species = function(x) "*", Petal.Width = function(x) "*")
kAnonData <- kAnon(iris, quasiIdentifiers = c("Species", "Petal.Width"), diversityFunctions, k = 3)
}
}
