% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{lambda}
\alias{lambda}
\alias{L}
\title{Lambda expressions}
\usage{
lambda(..., .env = parent.frame())

L(..., .env = parent.frame())
}
\arguments{
\item{...}{dots}

\item{.env}{environment of function}
}
\value{
function
}
\description{
Lambda expressions
}
