% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_div_subsets.r
\name{lDiversity}
\alias{lDiversity}
\title{Make a dataset l-diverse by applying anonymization functions}
\usage{
lDiversity(
  data,
  sensitiveAttributes,
  l,
  quasiIdentifiers = NULL,
  anonymizationFunctions = NULL,
  k = 5,
  shuffle = TRUE
)
}
\arguments{
\item{data}{The input dataset.}

\item{sensitiveAttributes}{A character vector specifying the names of the sensitive attributes.}

\item{l}{The desired minimum number of distinct values for each sensitive attribute within each group.}

\item{quasiIdentifiers}{A character vector specifying the names of the quasi-identifier columns (default: NULL).}

\item{anonymizationFunctions}{A named list of functions corresponding to the quasi-identifier columns.
Each function should take a vector as input and return a modified vector with the same length.}

\item{k}{(Default 5) The minimum number of rows in an l diverse subset. Not stricly an l-diversity
requirement, but here to accomodate for Finnish law / customs on
anonymous data publishing.}

\item{shuffle}{(Default TRUE) A parameter which determines whether the data
will be shuffled or not before returning. Will give warning
uf FALSE, as no data should be released without shuffling.
For diagnostic use, or for calculating reidentification rate.}
}
\value{
A dataset that is l-diverse with respect to the specified quasi-identifier columns and sensitive attributes,
  or an error is thrown if the desired level of l-diversity cannot be achieved.
}
\description{
This function takes a dataset and applies a set of diversity functions to the specified
quasi-identifier columns in order to achieve l-diversity with respect to the sensitive attributes.
It checks if the dataset is already l-diverse and returns the dataset as is in that case.
If not, it iteratively combines subsets of the dataset and applies the anonymity functions
until the desired level of l-diversity is achieved.
}
\examples{
\dontrun{
data <- data.frame(
  age = c(25, 30, 35, 40, 45),
  gender = c("M", "M", "F", "F", "M"),
  disease = c("A", "A", "B", "C", "C")
)

# Define functions for generalizing age and gender
age_fun <- function(x) floor(x / 10) * 10
gender_fun <- function(x) "*"

# Apply makeLdiverse function
ldiverse_data <- makeLdiverse(data, c("age", "gender"), "disease", list(age = age_fun, gender = gender_fun), 2)
}
}
