% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_anon_columnwise.r
\name{make_k_anonymous}
\alias{make_k_anonymous}
\title{Make a data frame k-anonymous by applying functions to quasi-identifier columns}
\usage{
make_k_anonymous(df, quasi_id_cols, fun_list, k)
}
\arguments{
\item{df}{The input data frame.}

\item{quasi_id_cols}{A character vector specifying the names of the quasi-identifier columns.}

\item{fun_list}{A named list of functions corresponding to the quasi-identifier columns.
Each function should take a vector as input and return a modified vector with the same length.}

\item{k}{The desired level of k-anonymity.}
}
\value{
A data frame that is k-anonymous with respect to the specified quasi-identifier columns,
  or an error is thrown if the desired level of k-anonymity cannot be achieved.
}
\description{
This function takes a data frame and applies a set of functions to the specified
quasi-identifier columns in order to achieve k-anonymity. It checks if the data frame
is already k-anonymous and returns a random permutation of the data in that case. If
not, it applies the functions iteratively to the columns until the desired level of
k-anonymity is achieved or it is not possible with the given functions.
}
\examples{
df <- data.frame(
  age = c(25, 30, 35, 40, 45),
  gender = c("M", "M", "F", "F", "M"),
  income = c(50000, 60000, 70000, 80000, 90000)
)

# Define functions to generalize age and income
age_fun <- function(x) floor(x / 10) * 10
income_fun <- function(x) round(x, -3)

# Apply make_k_anonymous function
k_anon_df <- make_k_anonymous(df, c("age", "income"), list(age = age_fun, income = income_fun), 2)

}
