% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_div_columnwise.r
\name{make_l_diverse}
\alias{make_l_diverse}
\title{Make a data frame l-diverse by applying functions to quasi-identifier and sensitive columns}
\usage{
make_l_diverse(df, quasi_id_cols, sensitive_cols, fun_list, l)
}
\arguments{
\item{df}{The input data frame.}

\item{quasi_id_cols}{A character vector specifying the names of the quasi-identifier columns.}

\item{sensitive_cols}{A character vector specifying the names of the sensitive columns.}

\item{fun_list}{A named list of functions corresponding to the quasi-identifier columns.
Each function should take a vector as input and return a modified vector with the same length.}

\item{l}{The desired minimum number of distinct values for each sensitive column within each group.}
}
\value{
A data frame that is l-diverse with respect to the specified quasi-identifier and sensitive columns,
  or an error is thrown if the desired level of l-diversity cannot be achieved.
}
\description{
This function takes a data frame and applies a set of functions to the specified
quasi-identifier columns in order to achieve l-diversity with respect to the sensitive columns.
It checks if the data frame is already l-diverse and returns a random permutation of the data in that case.
If not, it applies the functions iteratively to the columns until the desired level of l-diversity is achieved
or it is not possible with the given functions.
}
\examples{
df <- data.frame(
  age = c(25, 30, 35, 40, 45),
  gender = c("M", "M", "F", "F", "M"),
  disease = c("A", "A", "B", "C", "C")
)

# Define functions to generalize age and disease
age_fun <- function(x) floor(x / 10) * 10
gender_fun <- function(x) "*"

# Apply make_l_diverse function
l_diverse_df <- make_l_diverse(df, c("age", "gender"), "disease", list(age = age_fun, gender = gender_fun), 2)

}
