% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_div_subsets.r
\name{matrix_distance}
\alias{matrix_distance}
\title{importFrom caret dummyVars contr.ltfr
import caret
import MASS
Calculate the diversity distance between two subsets based on the given quasi-identifiers}
\usage{
matrix_distance(subset, otherSubset, quasiIdentifiers)
}
\arguments{
\item{subset}{A data frame representing one subset of the data.}

\item{otherSubset}{A data frame representing another subset of the data.}

\item{quasiIdentifiers}{A character vector containing the names of the columns that serve as quasi-identifiers.}
}
\value{
A numeric value representing the diversity distance between the two subsets.
}
\description{
This function calculates the diversity distance between two subsets based on the given quasi-identifiers.
}
\examples{
\donttest{
data1 <- list(data.frame(ID = 1:5, Age = c(25, 35, 40, 28, 32), Gender = c("M", "F", "F", "M", "F")))
data2 <- list(data.frame(ID = 6:10, Age = c(28, 38, 42, 31, 29), Gender = c("M", "F", "F", "F", "M")))
diversity_distance <- matrix_distance(data1, data2, c("Age", "Gender"))
}

}
