% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{meanDiffsinCor}
\alias{meanDiffsinCor}
\title{Mean Difference in Correlation Matrices}
\usage{
meanDiffsinCor(dataframelist)
}
\arguments{
\item{dataframelist}{A list of data frames to compare.}
}
\value{
A list containing two elements: "Medians" (median differences in correlation matrices) and "MADs" (median absolute deviations of correlation matrices).
}
\description{
Calculate the median difference in correlation matrices between data frames in a list, scaled by a reference data frame.
}
