% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{meanDiffsinVar}
\alias{meanDiffsinVar}
\title{Mean Difference in Column Variances}
\usage{
meanDiffsinVar(dataframelist)
}
\arguments{
\item{dataframelist}{A list of data frames to compare.}
}
\value{
A list containing two elements: "Means" (Mean differences in column variances) and "Sds" (mean standard deviations of column variances).
}
\description{
Calculate the mean difference in column variances between data frames in a list, scaled by a reference data frame.
}
