% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{medianDiffsinMean}
\alias{medianDiffsinMean}
\title{Median Difference in Column Means}
\usage{
medianDiffsinMean(dataframelist)
}
\arguments{
\item{dataframelist}{A list of data frames to compare.}
}
\value{
A list containing two elements: "Medians" (median differences in column means) and "MADs" (median absolute deviations of column means).
}
\description{
Calculate the median difference in column means between data frames in a list, scaled by a reference data frame.
}
