% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{medianDiffsinVar}
\alias{medianDiffsinVar}
\title{Median Difference in Column Variances}
\usage{
medianDiffsinVar(dataframelist)
}
\arguments{
\item{dataframelist}{A list of data frames to compare.}
}
\value{
A list containing two elements: "Medians" (median differences in column variances) and "MADs" (median absolute deviations of column variances).
}
\description{
Calculate the median difference in column variances between data frames in a list, scaled to a reference data frame.
}
