% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{minmax_scaler}
\alias{minmax_scaler}
\title{Min-Max scaling of a numeric vector.}
\usage{
minmax_scaler(column)
}
\arguments{
\item{column}{A numeric vector to be scaled.}
}
\value{
A scaled numeric vector with values in the range [0, 1].
}
\description{
This function performs Min-Max scaling on a numeric vector (column), scaling
its values to a range between 0 and 1. Min-Max scaling is a common technique
used to transform data to a specific range, preserving the relative differences
between values.
}
\examples{
column <- c(10, 20, 30, 40)
scaled_column <- minmax_scaler(column)
scaled_column
}
