% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{most_common}
\alias{most_common}
\title{Most Common Value}
\usage{
most_common(column)
}
\arguments{
\item{column}{The input column.}
}
\value{
A vector with the most common value repeated for the length of the input column.
}
\description{
Find the most common value in a column.
}
\examples{
data <- c(1, 2, 2, 3, 3, 3, 4, 4, 4, 4)
most_common(data)
}
