% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon_context.R
\name{new_anon_context}
\alias{new_anon_context}
\title{Constructor for anon_context objects}
\usage{
new_anon_context(x, ...)
}
\arguments{
\item{x}{The object to wrap with anon_context class}

\item{...}{Additional attributes to set}
}
\value{
Object with anon_context class
}
\description{
Constructor for anon_context objects
}
\keyword{internal}
