% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_auto.R
\name{nlp_auto}
\alias{nlp_auto}
\title{Generate NLP auto-anonymization options list}
\usage{
nlp_auto(
  default = FALSE,
  cardinal = default,
  date = default,
  event = default,
  fac = default,
  gpe = default,
  language = default,
  law = default,
  loc = default,
  money = default,
  norp = default,
  ordinal = default,
  org = default,
  percent = default,
  person = default,
  product = default,
  quantity = default,
  time = default,
  work_of_art = default,
  propn = default
)
}
\arguments{
\item{default}{Logical indicating the value to use for unset arguments. Default is FALSE}

\item{cardinal}{Logical indicating whether to automatically anonymize cardinal numbers}

\item{date}{Logical indicating whether to automatically anonymize dates}

\item{event}{Logical indicating whether to automatically anonymize events}

\item{fac}{Logical indicating whether to automatically anonymize facilities}

\item{gpe}{Logical indicating whether to automatically anonymize geopolitical entities}

\item{language}{Logical indicating whether to automatically anonymize languages}

\item{law}{Logical indicating whether to automatically anonymize laws}

\item{loc}{Logical indicating whether to automatically anonymize locations}

\item{money}{Logical indicating whether to automatically anonymize money}

\item{norp}{Logical indicating whether to automatically anonymize nationalities/religious/political groups}

\item{ordinal}{Logical indicating whether to automatically anonymize ordinal numbers}

\item{org}{Logical indicating whether to automatically anonymize organizations}

\item{percent}{Logical indicating whether to automatically anonymize percentages}

\item{person}{Logical indicating whether to automatically anonymize person names}

\item{product}{Logical indicating whether to automatically anonymize products}

\item{quantity}{Logical indicating whether to automatically anonymize quantities}

\item{time}{Logical indicating whether to automatically anonymize times}

\item{work_of_art}{Logical indicating whether to automatically anonymize works of art}

\item{propn}{Logical indicating whether to automatically anonymize proper nouns}
}
\value{
A named list of logical values that can be set with \code{\link[=options]{options()}}.
}
\description{
This function creates a named list of logical values indicating which types of
NLP entities should be automatically anonymized when using the core \code{\link[=anon]{anon()}}
function. It provides a centralized way to manage automatic NLP anonymization
settings when set as the global option \code{anon.nlp_auto}.
}
\examples{
# Get default auto-anonymization options (all TRUE)
default_auto <- nlp_auto(default = TRUE)

# Customize specific entity types (default, default is FALSE)
custom_auto <- nlp_auto(
  person = TRUE,
  org = TRUE
)

# Set as global option
previous_options <- options(anon.nlp_auto = custom_auto)

# Get current global option
options("anon.nlp_auto")

# Reset the previous global options
options(anon.nlp_auto = previous_options$anon.nlp_auto)

}
