% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_default_replacements.R
\name{nlp_default_replacements}
\alias{nlp_default_replacements}
\title{Generate default replacement options list for NLP entities}
\usage{
nlp_default_replacements(
  cardinal = "[CARDINAL]",
  date = "[DATE]",
  event = "[EVENT]",
  fac = "[FAC]",
  gpe = "[GPE]",
  language = "[LANGUAGE]",
  law = "[LAW]",
  loc = "[LOC]",
  money = "[MONEY]",
  norp = "[NORP]",
  ordinal = "[ORDINAL]",
  org = "[ORG]",
  percent = "[PERCENT]",
  person = "[PERSON]",
  product = "[PRODUCT]",
  quantity = "[QUANTITY]",
  time = "[TIME]",
  work_of_art = "[WORK_OF_ART]",
  propn = "[PROPN]"
)
}
\arguments{
\item{cardinal}{Default replacement for cardinal numbers (default: "[CARDINAL]")}

\item{date}{Default replacement for dates (default: "[DATE]")}

\item{event}{Default replacement for events (default: "[EVENT]")}

\item{fac}{Default replacement for facilities (default: "[FAC]")}

\item{gpe}{Default replacement for geopolitical entities (default: "[GPE]")}

\item{language}{Default replacement for languages (default: "[LANGUAGE]")}

\item{law}{Default replacement for laws (default: "[LAW]")}

\item{loc}{Default replacement for locations (default: "[LOC]")}

\item{money}{Default replacement for money (default: "[MONEY]")}

\item{norp}{Default replacement for nationalities/religious/political groups (default: "[NORP]")}

\item{ordinal}{Default replacement for ordinal numbers (default: "[ORDINAL]")}

\item{org}{Default replacement for organizations (default: "[ORG]")}

\item{percent}{Default replacement for percentages (default: "[PERCENT]")}

\item{person}{Default replacement for person names (default: "[PERSON]")}

\item{product}{Default replacement for products (default: "[PRODUCT]")}

\item{quantity}{Default replacement for quantities (default: "[QUANTITY]")}

\item{time}{Default replacement for times (default: "[TIME]")}

\item{work_of_art}{Default replacement for works of art (default: "[WORK_OF_ART]")}

\item{propn}{Default replacement for proper nouns (default: "[PROPN]"). This only applies
to \code{\link[=anon_nlp_proper_nouns]{anon_nlp_proper_nouns()}}.}
}
\value{
A named list of default replacement values that can be set with \code{\link[=options]{options()}}.
}
\description{
This function creates a named list of default replacement values for different
types of NLP entity-based anonymization. It provides a centralized way to manage
default replacements for entities anonymized in \code{\link[=anon_nlp_entities]{anon_nlp_entities()}} and
\code{\link[=anon_nlp_proper_nouns]{anon_nlp_proper_nouns()}} when set as the global option \code{anon.nlp_default_replacements}.
}
\examples{
# Get default options
default_opts <- nlp_default_replacements()

# Customize specific replacements
custom_opts <- nlp_default_replacements(
  person = "[NAME]",
  org = "[COMPANY]"
)

# Set as global option
previous_options <- options(anon.nlp_default_replacements = custom_opts)

# Get current global option
options("anon.nlp_default_replacements")

# Reset the previous global options
options(anon.nlp_default_replacements = previous_options$anon.nlp_default_replacements)

}
