% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon_measures.r
\name{prediction_all}
\alias{prediction_all}
\title{Prediction All Measures}
\usage{
prediction_all(original_data, k, reference_data, dist = euc_dist)
}
\arguments{
\item{original_data}{The original dataset.}

\item{k}{Number of kth-nearest records to consider.}

\item{reference_data}{The reference dataset.}

\item{dist}{Distance function to use (default: euc_dist).}
}
\value{
A list containing two sub-lists:
  - original_list: Contains the measures for the original_data without a reference_data.
  - reference_list: Contains the measures for the original_data with the provided reference_data.
}
\description{
Wrapper function to calculate all the measures for reference and original data,
store them into a list in the order of prediction distance,
prediction ambiguity, and prediction uncertainty,
and return the list.
}
\examples{
original_data <- as.data.frame(matrix(1:6, ncol = 2))
reference_data <- as.data.frame(matrix(7:12, ncol = 2))
prediction_all(original_data, k = 2, reference_data)

}
