% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon_measures.r
\name{prediction_ambiguity}
\alias{prediction_ambiguity}
\title{Prediction Ambiguity}
\usage{
prediction_ambiguity(original_data, k, reference_data = NULL, dist = euc_dist)
}
\arguments{
\item{original_data}{The original dataset for which prediction ambiguity needs to be calculated.}

\item{k}{The number of nearest neighbors to consider for calculating the ambiguity.}

\item{reference_data}{(Optional) The reference dataset used for calculating the distances. If not provided, the distances will be computed within the original_data.}

\item{dist}{The distance function to be used. Default is euclidean distance (euc_dist).}
}
\value{
A numeric vector representing the prediction ambiguity for each record in the original_data.
}
\description{
Calculates the prediction ambiguity by comparing the relative distance from each record to its nearest neighbor versus its kth nearest neighbor in the set of distances.
}
\examples{
prediction_ambiguity(original_data = iris, k = 3)


}
