% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon_measures.r
\name{prediction_distance}
\alias{prediction_distance}
\title{Prediction Distance}
\usage{
prediction_distance(original_data, reference_data = NULL, dist = euc_dist)
}
\arguments{
\item{original_data}{The original dataset.}

\item{reference_data}{The reference dataset (default: NULL).}

\item{dist}{Distance function to use (default: euc_dist).}
}
\value{
A numeric vector containing the distances between each row of the original_data and the reference_data.
}
\description{
Calculate distances between each row of the original_data and the reference_data using the specified distance function.
}
