% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon_measures.r
\name{prediction_plot}
\alias{prediction_plot}
\title{Prediction Plot}
\usage{
prediction_plot(prediction_all_output)
}
\arguments{
\item{prediction_all_output}{The output from the prediction_all function, containing the measures for original and reference data.}
}
\description{
Plot the measures in the original data against measures of the same type in the reference data.
}
\examples{
original_data <- as.data.frame(matrix(1:6, ncol = 2))
reference_data <- as.data.frame(matrix(7:12, ncol = 2))
preds = prediction_all(original_data, k = 2, reference_data)
prediction_plot(preds)

}
