% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy.r
\name{prediction_plot_list}
\alias{prediction_plot_list}
\title{Prediction Plot list version}
\usage{
prediction_plot_list(prediction_all_output, k, n = 10000, dist = euc_dist)
}
\arguments{
\item{prediction_all_output}{List of the prediction_all outputs for a single model with several repetitions}

\item{k}{The amount of neighbouring points to consider (recommended range: from 5 to 10)}

\item{n}{The number of points to plot (default: 1000)}

\item{dist}{The distance measure to use. (Defaults to eucledian distance)}
}
\description{
Plot the measures of predictive disclosure risk in an nonoverlapping sample against the measures of the same type in the anonymous data.
}
\examples{
\dontrun{
original_data <- as.data.frame(matrix(1:6, ncol = 2))
anon_data <- as.data.frame(matrix(7:12, ncol = 2))
prediction_plot(original_data, k = 2, anon_data)
}
}
