% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy.r
\name{prediction_uncertainty}
\alias{prediction_uncertainty}
\title{Prediction Uncertainty}
\usage{
prediction_uncertainty(
  original_data,
  k,
  anon_data = NULL,
  dist = euc_dist,
  distances = NULL
)
}
\arguments{
\item{original_data}{The original dataset.}

\item{k}{Number of best matches to find.}

\item{anon_data}{The reference dataset of the same column space (default: NULL).}

\item{dist}{Distance function to use (default: euc_dist).}

\item{distances}{precalculated distances (optional)}
}
\value{
A numeric vector containing the variances for each row of the original_data.
}
\description{
Prediction uncertainty gives the variation among the k best matches for each
row of the original data in the reference data for each row of the original
data, find the k best matches in the reference data using the dist function.
If the reference data is null, the function performs leave-one-out method to
calculate the distance between each row of the original data against the rest
nof the original data. Onehot encode and scale the datasets to
enable distance calculations.
}
