% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon_measures.r
\name{prediction_uncertainty_legacy}
\alias{prediction_uncertainty_legacy}
\title{Prediction Uncertainty}
\usage{
prediction_uncertainty_legacy(
  original_data,
  k,
  reference_data = NULL,
  dist = euc_dist
)
}
\arguments{
\item{original_data}{The original dataset.}

\item{k}{Number of best matches to find.}

\item{reference_data}{The reference dataset of the same column space (default: NULL).}

\item{dist}{Distance function to use (default: euc_dist).}
}
\value{
A numeric vector containing the variances for each row of the original_data.
}
\description{
Prediction uncertainty gives the variation among the k best matches for each row of the original data in the reference data for each row of the original data, find the k best matches in the reference data using the dist function. If the reference data is null, the function performs leave-one-out cross-validation to calculate the distance between each row of the original data against the rest of the original data. Onehot encode the datasets to enable distance calculations. Only run the numeric version if it is not null.
}
\examples{
prediction_uncertainty(original_data = iris, k = 3)

}
