% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon_context.R
\name{print.anon_context}
\alias{print.anon_context}
\title{Print method for single anon_context objects}
\usage{
\method{print}{anon_context}(x, ...)
}
\arguments{
\item{x}{An anon_context object}

\item{...}{Additional arguments passed to NextMethod}
}
\value{
Invisibly returns x
}
\description{
Print method for single anon_context objects
}
