% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon_fns.R
\name{print.fn_list}
\alias{print.fn_list}
\title{Print method for list of functions}
\usage{
\method{print}{fn_list}(x, ...)
}
\arguments{
\item{x}{A list of funtions with class \code{fn_list}}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for list of functions
}
