% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_bands.r
\name{quantile_bands}
\alias{quantile_bands}
\title{Divide a column into quantile bands}
\usage{
quantile_bands(column, n_bands, returnmean = FALSE, inclusive_tails = TRUE)
}
\arguments{
\item{column}{A numeric vector representing the column to be divided into bands.}

\item{n_bands}{An integer specifying the number of bands to create.}

\item{returnmean}{A logical value indicating whether to return the mean value of each band instead of the band itself.}

\item{inclusive_tails}{A logical value indicating whether to include the lowest and highest values in the first and last bands respectively.}
}
\value{
A character vector representing the quantile bands for each value in the column.
}
\description{
This function divides a column into quantile bands based on the specified number of bands.
}
\examples{
data <- c(10, 20, 15, 25, 30)
bands <- quantile_bands(data, 3, returnmean = FALSE, inclusive_tails = TRUE)
}
