% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon_measures.r
\name{reidentification_rate}
\alias{reidentification_rate}
\title{Calculate the reidentification rate of a dataset after anonymization}
\usage{
reidentification_rate(
  original_data,
  reference_data,
  quasiIdentifiers,
  dist = euc_dist
)
}
\arguments{
\item{original_data}{A data frame containing the original dataset}

\item{reference_data}{A data frame containing the reference dataset}

\item{quasiIdentifiers}{The quasi-identifiers of the data given as a vector of names.}

\item{dist}{An optional distance function to use for calculating distances between rows of the datasets. Defaults to the Euclidean distance function.}
}
\value{
A numeric value representing the proportion of correct guesses
}
\description{
This function calculates the reidentification rate of a dataset after it has been anonymized. The function one-hot encodes both the original and reference datasets to enable distance calculations, scales the datasets to have mean 0 and standard deviation 1, and then finds the nearest row in the reference data for each row of the original data. The function returns the proportion of correct guesses.
}
\examples{
original_data <- data.frame(x = c(1, 2, 3), y = c(4, 5, 6))
reference_data <- data.frame(x = c(2, 3, 4), y = c(5, 6, 7))
reidentification_rate(original_data, reference_data)
}
