% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy.r
\name{reidentification_rate_list}
\alias{reidentification_rate_list}
\title{Calculate the reidentification rate of a dataset after anonymization}
\usage{
reidentification_rate_list(
  data_list,
  quasiIdentifiers,
  n = 100,
  dist = euc_dist
)
}
\arguments{
\item{data_list}{A list containing original and anon items}

\item{quasiIdentifiers}{The quasi-identifiers of the data given as a vector of names.}

\item{n}{The number of times to calculate the reidentification rate}

\item{dist}{An optional distance function to use for calculating distances between rows of the datasets. Defaults to the Euclidean distance function.}
}
\value{
A numeric value representing the proportion of correct guesses
}
\description{
This function calculates the reidentification rate of a dataset after it has been anonymized. The function one-hot encodes both the original and reference datasets to enable distance calculations, scales the datasets to have mean 0 and standard deviation 1, and then finds the nearest row in the reference data for each row of the original data. The function returns the proportion of correct guesses.
}
\examples{
\dontrun{
reidentification_rate(data_list, quasiIdentifiers, n)
}
}
