% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump.R
\name{reidentification_rate_old}
\alias{reidentification_rate_old}
\title{Calculate the reidentification rate of a dataset after anonymization}
\usage{
reidentification_rate_old(
  original_data,
  anon_data,
  quasiIdentifiers,
  dist = euc_dist
)
}
\arguments{
\item{original_data}{A data frame containing the original dataset}

\item{anon_data}{A data frame containing the reference dataset}

\item{quasiIdentifiers}{The quasi-identifiers of the data given as a vector of names.}

\item{dist}{An optional distance function to use for calculating distances between rows of the datasets. Defaults to the Euclidean distance function.}
}
\value{
A numeric value representing the proportion of correct guesses
}
\description{
This function calculates the reidentification rate of a dataset after it has been anonymized. The function one-hot encodes both the original and reference datasets to enable distance calculations, scales the datasets to have mean 0 and standard deviation 1, and then finds the nearest row in the reference data for each row of the original data. The function returns the proportion of correct guesses.
}
