% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{reorder_rownames}
\alias{reorder_rownames}
\title{Reorder Rows Based on Row Names}
\usage{
reorder_rownames(df)
}
\arguments{
\item{df}{The input dataframe.}
}
\value{
A dataframe with rows reordered based on the extracted row numbers.
}
\description{
Reorder Rows Based on Row Names
}
