% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{roc_plot}
\alias{roc_plot}
\title{ROC Plot}
\usage{
roc_plot(model, model_anon, test, test_anon = NULL)
}
\arguments{
\item{model}{The model trained on original data.}

\item{model_anon}{The model trained on anonymized data.}

\item{test}{The test dataset where the response variable is on the first column.}

\item{test_anon}{The test dataset for anon data (if different from original data, mainly here for RSA)}
}
\description{
Generate an ROC plot comparing the performance of a model on original and anonymized data.
}
\examples{
\dontrun{
data("iris")
iris[,"Sepal.Length"] = ifelse(iris[,"Sepal.Length"] > 6, "1", "0")
rows = sample(c(TRUE, FALSE), replace = TRUE, nrow(iris), prob = c(0.7, 0.3))
train = iris[rows, ]
test = iris[!rows, ]
model = glm(as.factor(Sepal.Length) ~., data = train, family = binomial())
anon = spectral(train, cell_swap, shuffle = TRUE)
model_anon = glm(as.factor(Sepal.Length) ~., data = anon, family = binomial())
roc_plot(model, model_anon, test)
}

}
