% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{roc_plot_list}
\alias{roc_plot_list}
\title{ROC Plot list}
\usage{
roc_plot_list(
  targetIndex = 1,
  model_list,
  anon_model_list,
  test_list,
  test_anon_list = NULL
)
}
\arguments{
\item{targetIndex}{The index of the target variable in the test data.}

\item{model_list}{The model trained on original data.}

\item{anon_model_list}{The model trained on anonymized data.}

\item{test_list}{The test dataset where the response variable is on the first column.}

\item{test_anon_list}{The test dataset for anon data (if different from original data, mainly here for RSA)}
}
\description{
Generate an ROC plot comparing the performance of a model on original and anonymized data.
}
\examples{
\dontrun{
print("This is a test")
}

}
