% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{row_checker}
\alias{row_checker}
\title{Row Checker}
\usage{
row_checker(original_data, reference_data)
}
\arguments{
\item{original_data}{The original dataset.}

\item{reference_data}{The reference dataset.}
}
\description{
Check whether any row in the reference data is in the original data. If so, stop and raise an error.
Could be used to detect failure to anonymize data in some methods.
}
