% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{sensitive_generalizer}
\alias{sensitive_generalizer}
\title{Sensitive Column Generalizer}
\usage{
sensitive_generalizer(sensitiveColumn, subsetSize = NULL, subsets = NULL)
}
\arguments{
\item{sensitiveColumn}{The input sensitive column.}

\item{subsetSize}{The size of subsets to create.}

\item{subsets}{The number of subsets to create.}
}
\value{
A generalized version of the sensitive column with subsets replaced by their means (for numeric columns) or modes (for non-numeric columns).
}
\description{
Divide a sensitive column into subsets of a specified size or into a number of subsets and generalize each subset.
}
\examples{
numeric_data <- c(1, 2, 3, 4, 5, 6)
sensitive_generalizer(numeric_data, subsetSize = 2)

categorical_data <- c("A", "B", "B", "C", "C", "D")
sensitive_generalizer(categorical_data, subsetSize = 2)


}
