% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.r
\name{sensitive_noise}
\alias{sensitive_noise}
\title{Add Laplace-distributed noise to each column of the data.}
\usage{
sensitive_noise(data, epsilon = 1)
}
\arguments{
\item{data}{A data frame or matrix containing the original data.}

\item{epsilon}{The privacy parameter controlling the amount of noise added.}
}
\value{
A data frame or matrix with Laplace-distributed noise added to each column.
}
\description{
This function adds Laplace-distributed noise to each column of the input data.
The Laplace noise is controlled by the privacy parameter \code{epsilon}.
\code{epsilon} determines the scale
of the Laplace distribution. Larger \code{epsilon} means less noise.
}
\examples{
# Create a sample data frame
data <- data.frame(A = c(10, 12, 14), B = c(0, 5, 8))
# Add Laplace noise with epsilon = 0.1
sensitive_noise(data, epsilon = 0.1)

}
