% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{shuffle}
\alias{shuffle}
\title{Shuffle the values in each column of a data frame.}
\usage{
shuffle(data)
}
\arguments{
\item{data}{A data frame to be shuffled.}
}
\value{
A new data frame with the same dimensions as data, but with the
        values in each column randomly shuffled.
}
\description{
This function takes a data frame as input and returns a new data frame with
the same dimensions as the original data. The values in each column of the
returned data frame are shuffled randomly using the \code{sample} function,
effectively creating a random permutation of each column.
}
\examples{
data <- data.frame(x = c(1, 2, 3), y = c(4, 5, 6))
shuffled_data <- shuffle(data)
shuffled_data
}
