% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{signum}
\alias{signum}
\title{Significant Digits}
\usage{
signum(x, n)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{The number of significant digits to round to.}
}
\value{
A numeric vector rounded to n significant digits.
}
\description{
Significant Digits
}
\examples{
\dontrun{
a = 3.141
sig_num(a, 2)
}
}
