% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.r
\name{spectral}
\alias{spectral}
\title{Anonymize and decenter a data frame using spectral decomposition.}
\usage{
spectral(data, anonymizer)
}
\arguments{
\item{data}{A data frame to be anonymized and decentered.}

\item{anonymizer}{A function used to anonymize the U matrix from singluar
 value decomposition. The function should
take a data frame as input and return an anonymized data frame.
For example, the anonymizer function could perform a random
permutation of rows to achieve anonymization.}
}
\value{
A new data frame of the same size as data, containing the
        decentered and anonymized data.
}
\description{
This function takes a data frame as input and performs spectral decomposition
to anonymize the data by altering its principal components. The anonymization
is done by applying a user-defined anonymizer function to the principal
components. The function then reconstructs the decentered, anonymized data
and returns it as a new data frame.
}
