% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.r
\name{spectral_legacy}
\alias{spectral_legacy}
\title{Anonymize a data frame using spectral decomposition.}
\usage{
spectral_legacy(data, anonymizer, sample = FALSE, cat_as_num = FALSE)
}
\arguments{
\item{data}{A data frame to be anonymized and decentered.}

\item{anonymizer}{A function used to anonymize the U matrix from singluar
 value decomposition. The function should
take a data frame as input and return an anonymized data frame.
For example, the anonymizer function could perform a random
permutation of rows to achieve anonymization.}

\item{sample}{Logical: Whether to sample from the numerical columns
or take the max}

\item{cat_as_num}{Logical: Whether categorical variables should be returned
as numerical or not}
}
\value{
A new data frame containing the anonymized data.
}
\description{
This function takes a data frame as input and performs spectral decomposition
to anonymize the data by altering its left singular vectors. The anonymization
is done by applying a user-defined anonymizer function to vectors.
The function then reconstructs the decentered, anonymized data
and returns it as a new data frame.
}
