% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.r
\name{testRsquared_list}
\alias{testRsquared_list}
\title{Test R-squared for single model with several repetitions}
\usage{
testRsquared_list(model_list, test_datalist, responseVar)
}
\arguments{
\item{model_list}{A list of models trained on anony}

\item{responseVar}{The name of the response variable.}
}
\value{
A named list of R-squared values for each model.
}
\description{
Calculate R-squared values for multiple models on test data.
}
\examples{
\dontrun{
# Example usage:
model_list <- list(model1, model2, model3)
testRsquared(model_list, test_data, "response_variable")
}
}
