% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon_fns.R
\name{anon_phone_number}
\alias{anon_phone_number}
\title{Anonymize phone numbers}
\usage{
anon_phone_number(x)
}
\arguments{
\item{x}{A vector to anonymize}
}
\description{
Anonymizes phone numbers with sequential fake numbers.
}
