% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{combine_lowest_classes}
\alias{combine_lowest_classes}
\title{Combine the lowest classes}
\usage{
combine_lowest_classes(column, k = 1)
}
\arguments{
\item{column}{The column in which to combine the lowest classes.}

\item{k}{The maximum number of fewest occurances to combine.}
}
\value{
The column with the categorical values combined.
}
\description{
This function iteratively combines the fewest occuring categorical values in a column (k) times.
}
\examples{
column <- c("A", "B", "A", "C", "B", "D", "A", "C", "B")
combine_lowest_classes(column, 2)
}
