% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_div_subsets.r
\name{isLDiverse}
\alias{isLDiverse}
\title{Check if a dataset is l-diverse for a sensitive attribute}
\usage{
isLDiverse(data, sensitiveAttributes, quasiIdentifiers, l)
}
\arguments{
\item{data}{A data frame containing the sensitive data.}

\item{sensitiveAttributes}{A character vector specifying the column name of the sensitive attribute.}

\item{quasiIdentifiers}{A character vector specifying the column names of the quasi-identifiers.}

\item{l}{The desired level of l-diversity.}
}
\value{
A boolean indicating whether the dataset is l-diverse for the sensitive attribute.
}
\description{
This function checks if a dataset is l-diverse for a sensitive attribute
based on the quasi-identifiers and the desired level of l-diversity
}
\details{
The `isLDiverse` function checks if the input dataset `data`
is l-diverse for the sensitive attribute `sensitiveAttribute`
based on the quasi-identifiers specified in `quasiIdentifiers` and the desired level of l-diversity `l`.
}
