% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_anon_subsets.r
\name{kAnon}
\alias{kAnon}
\title{Perform k-Anonymization on a dataset}
\usage{
kAnon(
  data,
  k,
  quasiIdentifiers = NULL,
  anonymizationFunctions = NULL,
  shuffle = TRUE
)
}
\arguments{
\item{data}{A data frame containing the sensitive data.}

\item{k}{The desired level of k-anonymity.}

\item{quasiIdentifiers}{A character vector specifying the column names of the quasi-identifiers (default: NULL).}

\item{anonymizationFunctions}{A named list of anonymization functions for each quasi-identifier column (default: NULL).}

\item{shuffle}{Whether to shuffle the dataset before returning. Warning if FALSE, used to calculate empirical reidentification rate.}
}
\value{
A k-anonymous data frame.
}
\description{
This function applies k-anonymization to a dataset by dividing it into subsets
based on quasi-identifiers and modifying the data to achieve k-anonymity
}
\details{
The `kAnon` function applies k-anonymization on the input dataset `data`
by dividing it into subsets based on quasi-identifiers specified in `quasiIdentifiers`.
The function modifies the data in each subset to achieve k-anonymity
based on the anonymization functions provided in `anonymizationFunctions`.
}
\examples{
\dontrun{
data(iris)
anonymizationFunctions <- list(Species = function(x) "*", Petal.Width = function(x) "*")
kAnonData <- kAnon(iris, quasiIdentifiers = c("Species", "Petal.Width"), anonymizationFunctions, k = 3)
}
}
