% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_div_subsets.r
\name{lDiversity}
\alias{lDiversity}
\title{Make a dataset l-diverse by applying (user provided or default) anonymization functions to a set of quasi-identifiers.}
\usage{
lDiversity(
  data,
  sensitiveAttributes,
  l,
  quasiIdentifiers = NULL,
  anonymizationFunctions = NULL,
  k = 1,
  shuffle = FALSE
)
}
\arguments{
\item{data}{The input dataset.}

\item{sensitiveAttributes}{A character vector specifying the names of the sensitive attributes.}

\item{l}{The desired minimum number of distinct values for each sensitive attribute within each group.}

\item{quasiIdentifiers}{A character vector specifying the column names of the quasi-identifiers (default: NULL). The first column named will be modified first and the last last:
name the columns in ascending order of importance. If quasi-identifiers are nor provided, all columns will be assumed to be quasi-identifiers and modified in descending order of cardinality.}

\item{anonymizationFunctions}{A named list of anonymization functions for each quasi-identifier column (default: NULL). If not provided, mean and mode will be used for numericals and non-numericals respectively.}

\item{k}{(Default 1) The minimum number of rows in an l diverse subset. Not stricly an l-diversity
requirement, but here to be used if needed. Can be increased to improve privacy.}

\item{shuffle}{Whether to shuffle the dataset before returning. Warning if FALSE, used to calculate empirical reidentification rate.}
}
\value{
An l-diverse data frame.
}
\description{
Make a dataset l-diverse by applying (user provided or default) anonymization functions to a set of quasi-identifiers.
}
\details{
This function takes a dataset and applies a set of anonymity functions to the specified
quasi-identifier columns in order to achieve l-diversity with respect to the sensitive attributes.
It iteratively combines subsets of the dataset and applies the anonymity functions
until the desired level of l-diversity is achieved. Quasi-identifiers are
columns which combined can identify a person in the data. Quasi-identifiers
include but are not limited to: age, gender, zip-code, profession...
Sensitive attributes are attributes which values could cause harm to a person
if revealed. Sensitive attributes may include: health information and beliefs etc.
}
\examples{
\dontrun{
data("iris")
lDiversity(iris, "Species", l = 2)
}
}
