% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_div_subsets.r
\name{matrix_distance}
\alias{matrix_distance}
\title{Calculate the diversity distance between two subsets based on the given quasi-identifiers}
\usage{
matrix_distance(subset, otherSubset, quasiIdentifiers)
}
\arguments{
\item{subset}{A data frame representing one subset of the data.}

\item{otherSubset}{A data frame representing another subset of the data.}

\item{quasiIdentifiers}{A character vector containing the names of the columns that serve as quasi-identifiers.}
}
\value{
A numeric value representing the diversity distance between the two subsets.
}
\description{
This function calculates the diversity distance between two subsets based on the given quasi-identifiers.
}
