% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more_patterns.R
\name{more_patterns}
\alias{more_patterns}
\title{Create expanded pattern vector for text matching}
\usage{
more_patterns(
  patterns,
  original = TRUE,
  escape_regex = TRUE,
  spaces_to_flexible = TRUE,
  individual = TRUE
)
}
\arguments{
\item{patterns}{Character vector of patterns to expand}

\item{original}{Logical. Whether to include original patterns (default: TRUE)}

\item{escape_regex}{Logical. Whether to include original patterns with special characters escaped (default: TRUE)}

\item{spaces_to_flexible}{Logical. Whether to include space-flexible patterns (default: TRUE)}

\item{individual}{Logical. Whether to include individual words (default: TRUE)}
}
\value{
Character vector with selected pattern expansions
}
\description{
Accepts a character vector of patterns and returns an expanded vector containing:
\enumerate{
\item Original items
\item Original items with special characters escaped
\item Patterns with spaces replaced with a pattern of any number of spaces and any
one character between words.
\item Individual words from each item, split by spaces and punctuation, not including stop words.
}
}
\details{
The intention of the order is to replace phrases with as few replacements as possible
while taking additional efforts to anonymize all sensitive information.
}
\examples{
people <- c("John Smith", "Mary Jane Watson")
more_patterns(people)
more_patterns(people, individual = FALSE)
}
