% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy.r
\name{prediction_ambiguity}
\alias{prediction_ambiguity}
\title{Prediction Ambiguity}
\usage{
prediction_ambiguity(
  original_data,
  k,
  anon_data = NULL,
  dist = euc_dist,
  distances = NULL
)
}
\arguments{
\item{original_data}{The original dataset for which prediction ambiguity needs to be calculated.}

\item{k}{The number of nearest neighbors to consider for calculating the ambiguity.}

\item{anon_data}{(Optional) The reference dataset used for calculating the distances. If not provided, the distances will be computed within the original_data.}

\item{dist}{The distance function to be used. Default is euclidean distance (euc_dist).}

\item{distances}{precalculated distances (optional)}
}
\value{
A numeric vector representing the prediction ambiguity for each record in the original_data.
}
\description{
Calculates the prediction ambiguity by comparing the relative distance from each record to its nearest neighbor versus its kth nearest neighbor in the set of distances.
}
