% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{shuffle}
\alias{shuffle}
\title{Shuffle the rows of a data frame.}
\usage{
shuffle(data)
}
\arguments{
\item{data}{A data frame to be shuffled.}
}
\value{
A new data frame with the same dimensions as data, but with the
        rows shuffled.
}
\description{
This function takes a data frame as input and returns a new data frame with
the same dimensions as the original data. The rows of the
returned data frame are shuffled randomly using the \code{sample} function,
effectively re-ordering the dataframe, as in shuffling a deck of cards.
}
\examples{
data <- data.frame(x = c(1, 2, 3), y = c(4, 5, 6))
shuffled_data <- shuffle(data)
shuffled_data
}
