% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.r
\name{spectral}
\alias{spectral}
\title{Anonymize a data frame using spectral anonymization.}
\usage{
spectral(
  data,
  anonymizer,
  on_matrices = "U",
  preserveMeans = TRUE,
  full = FALSE,
  sample = FALSE,
  cat_as_num = FALSE,
  shuffle = FALSE
)
}
\arguments{
\item{data}{A data frame to be anonymized and decentered.}

\item{anonymizer}{A function used to anonymize part of the singluar
 value decomposition. The function should
take a data frame as input and return an anonymized data frame of the same size.}

\item{on_matrices}{A character string indicating which matrix to anonymize.
Possible values are "U", "V", "UD", and "DV".
The default is "U". Note that the transposition is not
shown here on V for ease of parametrization, but the
modifications will be made on it's transpose, V'.}

\item{preserveMeans}{Extract colmeans before svd or not, and add them back afte svd anoymization. Use for cell swap at least, might be usefull for other functions as well. Defaults TRUE.}

\item{full}{Logical: Whether to use the full SVD or the reduced SVD (default).}

\item{sample}{Logical: Whether to sample a value for a categorical variable represented in the probabilistic state or take the value with the most weight.}

\item{cat_as_num}{Logical: Whether categorical variables should be returned
in their probabilistic state or not.}

\item{shuffle}{Logical: Whether to shuffle the dataset before returning. Warning if FALSE, used to calculate empirical reidentification rate.}
}
\value{
A new data frame containing the anonymized data.
}
\description{
This function takes a data frame M as input and performs singular value decomposition on it, M = UDV'.
Then the user defined anonymization function is performed on some matrix (or matrices) of the decomposition.
M is reconstructed using the anonymized SVD and returned. Built in functions that work with spectral are
"cell_swap" and "sensitive_noise".
}
